make c-stdlib file names more consistent
