+++ lualib/c-stdlib/my_math.h
@@ -1,8 +1,3 @@
-#ifndef LUA_C_STDLIB_MATH_H_
-#define LUA_C_STDLIB_MATH_H_
-
-#define DBL_MAX 1.79769313486231570815e+308
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -15,49 +10,78 @@ extern "C" {
 
 #ifndef fp_force_evalf
 #define fp_force_evalf fp_force_evalf
-static inline void fp_force_evalf(float x)
-{
-	volatile float y;
-	y = x;
-	(void)y;
+static inline void fp_force_evalf(float x) {
+    volatile float y;
+    y = x;
+    (void)y;
 }
 #endif
 
 #ifndef fp_force_eval
 #define fp_force_eval fp_force_eval
-static inline void fp_force_eval(double x)
-{
-	volatile double y;
-	y = x;
-	(void)y;
+static inline void fp_force_eval(double x) {
+    volatile double y;
+    y = x;
+    (void)y;
 }
 #endif
 
 #ifndef fp_force_evall
 #define fp_force_evall fp_force_evall
-static inline void fp_force_evall(long double x)
-{
-	volatile long double y;
-	y = x;
-	(void)y;
+static inline void fp_force_evall(long double x) {
+    volatile long double y;
+    y = x;
+    (void)y;
 }
 #endif
 
-#define FORCE_EVAL(x) do {                        \
-	if (sizeof(x) == sizeof(float)) {         \
-		fp_force_evalf(x);                \
-	} else if (sizeof(x) == sizeof(double)) { \
-		fp_force_eval(x);                 \
-	} else {                                  \
-		fp_force_evall(x);                \
-	}                                         \
-} while(0)
+#define FORCE_EVAL(x)                             \
+    do {                                          \
+        if (sizeof(x) == sizeof(float)) {         \
+            fp_force_evalf(x);                    \
+        } else if (sizeof(x) == sizeof(double)) { \
+            fp_force_eval(x);                     \
+        } else {                                  \
+            fp_force_evall(x);                    \
+        }                                         \
+    } while (0)
 
-double      fmod(double, double);
-double      frexp(double, int *);
+double acos(double);
+
+double asin(double);
+
+double atan2(double, double);
+
+double cos(double);
+
+double cosh(double);
+
+double exp(double);
+
+double fabs(double);
+
+double log(double);
+
+double log2(double);
+
+double log10(double);
+
+double pow(double, double);
+
+double sin(double);
+
+double sinh(double);
+
+double sqrt(double);
+
+double tan(double);
+
+double tanh(double);
+
+double scalbn(double, int);
+
+double ldexp(double, int);
 
 #ifdef __cplusplus
 }
-#endif
-
 #endif
