+++ lualib/c-stdlib/src/mocked_math_impl.h
@@ -1,4 +1,4 @@
-#include "../mocked_math.h"
+#include "../my_math.h"
 #include <stdint.h>
 
 #include "math_pow.c"
