+++ lualib/c-stdlib/src/mocked_stdio_impl.h
@@ -1,4 +1,4 @@
-#include "../mocked_stdio.h"
+#include "../my_stdio.h"
 
 #include <stdlib.h>
 #include <string.h>
