+++ lualib/lauxlib.c
@@ -11,7 +11,7 @@
 #include <stdarg.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lprefix.h"
