+++ lualib/liolib.c
@@ -12,7 +12,7 @@
 #include <locale.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lauxlib.h"
 #include "lprefix.h"
