+++ lualib/loadlib.c
@@ -13,7 +13,7 @@
 
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lauxlib.h"
 #include "lprefix.h"
