+++ lualib/lstrlib.c
@@ -15,7 +15,7 @@
 #include <stddef.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lauxlib.h"
 #include "lprefix.h"
