+++ lualib/lua.c
@@ -11,7 +11,7 @@
 #include <signal.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lauxlib.h"
 #include "lprefix.h"
