+++ lualib/luac.c
@@ -11,7 +11,7 @@
 #include <errno.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "lauxlib.h"
 #include "ldebug.h"
