+++ lualib/lvm.c
@@ -14,7 +14,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include <string.h>
-#include "mocked_stdio.h"
+#include "my_stdio.h"
 
 #include "ldebug.h"
 #include "ldo.h"
